Aassignment 2 - November 14, 2018
Here's the required output for the assignment no 1
This assignment was locked Nov 7 at 5pm.
1. WHO dataset

d. country with the lowest literacy
e. Richest country in Europe based on GNI
f. Mean Life expectancy of countries in Africa
g. Number of countries with population greater than 10,000
h. Top 5 countries in the Americas with the highest child mortality

2. NBA dataset (Historical NBA Performance.xlsx)

a. The year Bulls has the highest winning percentage 
b. Teams with an even win-loss record in a year

3. Seasons_Stats.csv

a. Player with the highest 3-pt attempt rate in a season. 
b. Player with the highest free throw rate in a season. 
c. What year/season does Lebron James scored the highest? 
d. What year/season does Michael Jordan scored the highest? 
e. Player efficiency rating of Kobe Bryant in the year where his MP is the lowest?

4. National Universities Rankings.csv

a. University with the most number of undergrads 
b. Average Tuition in the Top 10 University